requirements.txt
Flask
mercadopago
gunicorn
Pillow
